+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Features:
   * rename `_Logger` class to `BaseLogger` 
   * move common info/warning/error/debug methods to BaseLogger class
   * `propagate = False` is now false for all loggers
+* `libs.py.helpers` create_dir, create_file use CliLogger for output messages instead of print 
 
 # v0.8.1